support non 32-bit length opcodes (#89)
* support non 32-bit length opcodes

this retrofits the eval and dis modules to use a bigint for the opcode
during the decode phase, supporting pcodes of arbitrary length and
(potentially) variable-length opcodes.

once a particular encoding is identified through the decode tree,
we convert the bigint to a fixed-width bitvector. this is then
passed through to the rest of the evaluation / analysis.

* add best-effort error when given opcode is too long

this is not robust, as it only triggers if non-zero bits are set
above the expected length. after parsing into a bigint,
leading zeros become indiscernible.