Scala backend & lifter-optimiser bdd copyprop (#102)
* Add scala backend for BASIL

* Add lifttime PC variable

Optional support for offline lifter to specify the PC value
at lifttime.

* Remove problematic list, add missing variables

* Optional pass to remove unsupported globals

* Whitespace in transforms

* Re-enable case simp for offline

* Decoder cleanup and sanity checks

* Simplify reachability based on enc

* Add lifttime PC variable

Optional support for offline lifter to specify the PC value
at lifttime.

* Remove problematic list, add missing variables

* Optional pass to remove unsupported globals

* Whitespace in transforms

* Re-enable case simp for offline

* Decoder cleanup and sanity checks

* Simplify reachability based on enc

* wip rt copyprop

* use bdd lattice for clobbered & read

* refac

* untested xform to ternary

* not working

* refactor so analysis and transform both called into from bdd AI walk

* fix anlaysis a bit and add bvadd

* cleanup bvops

* Eliminate comparisons during IntToBits

Leverage interval information to reduce trivial comparisons
during post-passes, then cleanup in RemoveUnused.

* fix rebuild expr

* passing cov disabled

* test and fix bdd sle/slt

* Merged backends (#80)

* Add scala backend for BASIL

* Add C++ Backend for LLVM 

* add aslBackwardsVisitor

* aslVisitor: change vstmt to return stmt list

BREAKING!
This change affects the signature of the Asl_visitor.visit_stmt method.
For compatibility, a visit_stmt_single method is provided with
equivalent behaviour to the old visit_stmt. There is also an added
helper function to convert visitActions on single statements to
visitActions on a list of statements.

Both of these compatibility helpers WILL THROW if used with a visitor
that returns non-singleton statement lists.

This gives the user the flexibility to insert new statements or delete
statements entirely. On the other hand, post-functions in
ChangeDoChildrenPost will need to handle lists of functions as well.

This follows the original CIL visitor: https://people.eecs.berkeley.edu/~necula/cil/api/Cil.cilVisitor.html

* fix backwards visitor and rearrange code

it is no longer a good idea for the backwards and forwards visitors to have a subtyping relation.

* support -x 0 to print encoding name. (#78)

this is very useful when looking for the name of an encoding,
without cluttering the output with the disassembly trace.

the default debug_level has been lowered to -1 to support -x 0
as a non-default level. we cannot print by default since that
would clutter stdout when used as a library.



Co-authored-by: rina <k@rina.fyi>

* progress

* possibly working

* build standalone

* print lifted semantics

* readme

* add mill script

* add generic lifter interface

* delete old utils

* generate separate scala assembly file

* cleanup

* update tests

* marshall offline lifter

* fix symbolic lifter merge

* update action

* update cpp lifter

* cpp: allow llvm 18

* cpp: clean up offlineASL-cpp folder

delete old hpp files in root

reset meson.build to empty

---------

Co-authored-by: Nicholas Coughlin <n.coughlin@uq.edu.au>
Co-authored-by: rina <k@rina.fyi>