fix symbolic field handling by `CommonSubExprElim` (#5)
Fix a problem with the treatment of symbolic field variables by
common-subexpression-elimination.

When a common subexpression has been identified, it needs to place it in
the statement list. It does this by placing it in the first position
after all its dependencies have been assigned. The problem comes because
a common subexpression that depends on a symbolic field won't ever find
an assignment to the symbolic field.

This is what the algorithm was probably supposed to do, but as they
explain in a comment what they actually do is place the definition of
the common subexpression in the first place after all its dependencies
have _appeared_ (whether assigned to or not). Therefore in practice,
common expressions involving symbolic fields would end up placed after
their first use.

This PR fixes it by:

* Passing a list of globals into CSE which are treated as predefined
* Appending symbolic fields to the global identifier set