first pass at a BNFC grammar for the rASL AST format (#143)
(the AST format is the one that looks like this: Stmt_ConstDecl(Type_Bits(64),"Cse0__5",Expr_TApply("add_bits.0",... and is used in gtirb-semantics and BASIL)

This adds a BNFC grammar to gain all the benefits of BNFC, most importantly, multi-language parser generator generating. This is the same language which is described in Semantics.g4, but BNFC brings some improvements and I've also just written a better grammar this time.

Some notes:

    I've tried to be reasonably precise in the grammar. For instance, there is a fixed list of intrinsic operations and variable names, and narrower rules are used where possible. Much of this is taken from the IBI for the offline lifter.
    The bnfc-generated files are checked in to the repo using Dune's "promote" feature. This means that bnfc is not a build dependency, as most people can just use the checked-in files. Only if you want to update the BNFC grammar and re-generate the files will BNFC be needed. Then, it will search for bnfc in PATH. (I didn't really want to write Dune code to download the bnfc binary).

Tested by running through all the AST outputs in test_cntlm.t and also by adding a BNFC parse to the coverage checking. This passes with no problems, so we can be sure that the BNFC grammar describes a superset of the AST which we see in coverage.

Commits:

* starting bnfc

* add statements

* add lots of prim names from offline IBI

* add .0 suffixes ;-;

* fix bugs and add StmtLines for newline-separated

* starting to do coverage

* dune promote for bnfc files

* invoke bnfc-generated things directly in coverage test

* mark as generated

* remove unconstrained tycon case

* add TestSemantics exe, move libASL_bnfc into libASL-stage1

* review: remove makefile, remove unix, use public asli.libASL-bnfc name